# TP2/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MythicGames REST API",
    "version": "1.0.0",
    "description": "API REST pour la marketplace de jeux vidéo MythicGames (TP2 - NodeJS + PostgreSQL)"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Serveur local" }
  ],
...
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Lister tous les produits",
        "parameters": [
          { "name": "name",  "in": "query", "schema": { "type": "string" }, "description": "Filtrer par nom" },
          { "name": "about", "in": "query", "schema": { "type": "string" }, "description": "Filtrer par description" },
          { "name": "price", "in": "query", "schema": { "type": "number" }, "description": "Filtrer par prix exact" }
        ],
        "responses": { "200": { "description": "Liste des produits" } }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Créer un produit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Produit créé" },
          "400": { "description": "Données invalides" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Récupérer un produit",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Produit trouvé" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Remplacer un produit (PUT)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductInput" } } }
        },
        "responses": { "200": { "description": "Produit mis à jour" }, "404": { "description": "Not found" } }
      },
      "patch": {
        "tags": ["Products"],
        "summary": "Mettre à jour partiellement un produit (PATCH)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductPatch" } } }
        },
        "responses": { "200": { "description": "Produit mis à jour" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Supprimer un produit",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Produit supprimé" }, "404": { "description": "Not found" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProductInput": {
        "type": "object",
        "required": ["name", "about", "price"],
        "properties": {
          "name":  { "type": "string", "example": "MythicGame Alpha" },
          "about": { "type": "string", "example": "Un RPG épique" },
          "price": { "type": "number", "example": 59.99 }
        }
      },
      "ProductPatch": {
        "type": "object",
        "properties": {
          "name":  { "type": "string" },
          "about": { "type": "string" },
          "price": { "type": "number" }
        }
      },
...
    }
  }
}